# expts: assess performance of structured outputs