@shreyashankar 2025-01-29
I will merge this script because I'm sure we will use it in the future